refactor: remove Python dependency and add cross-platform builds
Remove all Python/PyO3 dependencies (pyo3, serde-pyobject, include_dir),
the Python MCP module, and Python setup from CI, Docker, and the binary
entrypoint. The update_mcp_endpoint handler (which only uses WriteSide)
is preserved in admin.rs. Replace the single-platform release workflow
with a matrix strategy building for linux-x86_64, macos-arm64,
macos-x86_64, and windows-x86_64 (3 binary variants each). Simplify
Dockerfiles to two-stage Rust-only builds with debian:bookworm-slim
runtime. Fix flaky vector search test by adding similarity thresholds.
