+++ .github/workflows/release-build.yml
@@ -8,91 +8,86 @@ permissions:
   contents: write
 
 jobs:
-  build-and-patch-release:
-    name: Build and Upload Release Assets
-    runs-on: [self-hosted, X64]
+  build-native-binaries:
+    name: Build ${{ matrix.platform }}
+    runs-on: ${{ matrix.runner }}
+    strategy:
+      fail-fast: false
+      matrix:
+        include:
+          - platform: linux-x86_64
+            runner: [self-hosted, X64]
+            target: x86_64-unknown-linux-gnu
+            ext: ""
+          - platform: macos-arm64
+            runner: macos-latest
+            target: aarch64-apple-darwin
+            ext: ""
+          - platform: macos-x86_64
+            runner: macos-13
+            target: x86_64-apple-darwin
+            ext: ""
+          - platform: windows-x86_64
+            runner: windows-latest
+            target: x86_64-pc-windows-msvc
+            ext: ".exe"
+
+    env:
+      RUSTY_V8_MIRROR: https://github.com/denoland/rusty_v8/releases/download
+      V8_FROM_SOURCE: "0"
+      GIT_COMMIT: ${{ github.sha }}
+      CARGO_PROFILE_RELEASE_STRIP: debuginfo
 
     steps:
       - name: Checkout code
         uses: actions/checkout@v4
 
-      - name: Install dependencies
+      - name: Install system dependencies (Linux)
+        if: runner.os == 'Linux'
         run: |
           sudo apt-get update
-          sudo apt-get install -y protobuf-compiler python3 python3-dev python3-venv python3-pip libssl-dev pkg-config g++ build-essential ca-certificates librocksdb-dev libgflags-dev libsnappy-dev zlib1g-dev libbz2-dev libzstd-dev libxml2-dev
-
-      - name: Install Python dependencies
-        run: |
-          python3 -m venv .venv
-          source .venv/bin/activate
-          pip install --upgrade pip
-          pip install -r requirements.txt
+          sudo apt-get install -y libssl-dev pkg-config g++ build-essential ca-certificates
 
       - name: Setup Rust
-        uses: actions-rs/toolchain@v1
+        uses: dtolnay/rust-toolchain@master
         with:
-          profile: minimal
           toolchain: 1.88.0
-          override: true
+          targets: ${{ matrix.target }}
 
       - name: Cache Rust dependencies
         uses: Swatinem/rust-cache@v2
         with:
           cache-all-crates: "true"
 
-      - name: Build Default Binary
-        env:
-          PYO3_PYTHON: python3
-          RUSTY_V8_MIRROR: https://github.com/denoland/rusty_v8/releases/download
-          V8_FROM_SOURCE: "0"
-          GIT_COMMIT: ${{ github.sha }}
-        uses: actions-rs/cargo@v1
-        with:
-          command: build
-          args: --release
+      - name: Build compact binary
+        shell: bash
+        run: |
+          cargo build --release --target ${{ matrix.target }}
+          cp ./target/${{ matrix.target }}/release/oramacore${{ matrix.ext }} ./target/${{ matrix.target }}/release/oramacore-${{ matrix.platform }}-compact${{ matrix.ext }}
 
-      - name: Save Default Binary
-        run: cp ./target/release/oramacore ./target/release/oramacore-compact
+      - name: Build reader-only binary
+        shell: bash
+        run: |
+          cargo build --release --target ${{ matrix.target }} --no-default-features --features reader --bin oramacore
+          cp ./target/${{ matrix.target }}/release/oramacore${{ matrix.ext }} ./target/${{ matrix.target }}/release/oramacore-${{ matrix.platform }}-reader${{ matrix.ext }}
 
-      - name: Build Reader Binary
-        env:
-          PYO3_PYTHON: python3
-          RUSTY_V8_MIRROR: https://github.com/denoland/rusty_v8/releases/download
-          V8_FROM_SOURCE: "0"
-          GIT_COMMIT: ${{ github.sha }}
-        uses: actions-rs/cargo@v1
-        with:
-          command: build
-          args: --no-default-features --release --features reader --bin oramacore
+      - name: Build writer-only binary
+        shell: bash
+        run: |
+          cargo build --release --target ${{ matrix.target }} --no-default-features --features writer --bin oramacore
+          cp ./target/${{ matrix.target }}/release/oramacore${{ matrix.ext }} ./target/${{ matrix.target }}/release/oramacore-${{ matrix.platform }}-writer${{ matrix.ext }}
 
-      - name: Save Reader Binary
-        run: cp ./target/release/oramacore ./target/release/oramacore-reader
-
-      - name: Build Writer Binary
-        env:
-          PYO3_PYTHON: python3
-          RUSTY_V8_MIRROR: https://github.com/denoland/rusty_v8/releases/download
-          V8_FROM_SOURCE: "0"
-          GIT_COMMIT: ${{ github.sha }}
-        uses: actions-rs/cargo@v1
-        with:
-          command: build
-          args: --no-default-features --release --features writer --bin oramacore
-
-      - name: Save Writer Binary
-        run: cp ./target/release/oramacore ./target/release/oramacore-writer
-
-      - name: Upload Binaries to Release
+      - name: Upload binaries to Release
         uses: softprops/action-gh-release@v1
         with:
           files: |
-            ./target/release/oramacore-compact
-            ./target/release/oramacore-reader
-            ./target/release/oramacore-writer
+            ./target/${{ matrix.target }}/release/oramacore-${{ matrix.platform }}-compact${{ matrix.ext }}
+            ./target/${{ matrix.target }}/release/oramacore-${{ matrix.platform }}-reader${{ matrix.ext }}
+            ./target/${{ matrix.target }}/release/oramacore-${{ matrix.platform }}-writer${{ matrix.ext }}
           token: ${{ secrets.GITHUB_TOKEN }}
 
   build-docker-OramaCore-x86_64:
-    needs: [build-and-patch-release]
+    needs: [build-native-binaries]
     name: Build and Upload Docker Images
     runs-on: [self-hosted, X64]
 
@@ -147,7 +142,7 @@ jobs:
             oramasearch/oramacore:${{ steps.get_version.outputs.VERSION }}
 
   build-docker-OramaCore-arm64:
-    needs: [build-and-patch-release]
+    needs: [build-native-binaries]
     name: Build and Upload Docker Images
     runs-on: [self-hosted, arm64]
