+++ Cargo.toml
@@ -92,9 +92,6 @@ jwt-authorizer = "0.15"
 jsonwebtoken = "9.3.1"
 schemars = { version = "1.0.4", features = ["chrono04"] }
 http = "1.3.1"
-pyo3 = { version = "0.27.1" }
-serde-pyobject = "0.7.0"
-include_dir = "0.7"
 derive_more = "2.0.1"
 base64 = "0.22"
 fastembed = { version = "4.4.0", features = ["ort-download-binaries"] }
