+++ docker/Dockerfile-oramacore-arm64
@@ -1,19 +1,7 @@
-# Stage 1: Build Python environment with dependencies
-FROM python:3.11-slim-bookworm AS python-builder
-
-WORKDIR /app
-
-# Install Python dependencies into a virtual environment
-# Use CPU-only requirements file (compatible with ARM64, no CUDA packages)
-COPY requirements-cpu.txt .
-RUN python3.11 -m venv .venv && \
-    .venv/bin/pip install --upgrade pip && \
-    .venv/bin/pip install -r requirements-cpu.txt
-
-# Stage 2: Build Rust application with PyO3
+# Stage 1: Build Rust application
 FROM rust:1.85-slim-bookworm AS rust-builder
 
-# Install build dependencies including Python 3.11 dev headers for PyO3
+# Install build dependencies
 RUN apt-get update && apt-get install -y \
   libssl3 \
   pkg-config \
@@ -21,15 +9,6 @@ RUN apt-get update && apt-get install -y \
   g++ \
   build-essential \
   ca-certificates \
-  libgflags-dev \
-  libsnappy-dev \
-  zlib1g-dev \
-  libbz2-dev \
-  libzstd-dev \
-  libxml2-dev \
-  protobuf-compiler \
-  python3.11 \
-  python3.11-dev \
   curl \
   git \
   && rm -rf /var/lib/apt/lists/*
@@ -39,23 +18,14 @@ ENV V8_FROM_SOURCE=0
 
 WORKDIR /usr/src/app
 
-# Copy Python venv from python-builder stage (for runtime dependencies)
-COPY --from=python-builder /app/.venv /app/.venv
-
-# Set Python environment for PyO3 to use system Python (not venv)
-# The venv Python interpreter is just a symlink that breaks across stages
-ENV PYO3_PYTHON=/usr/bin/python3.11
-ENV VIRTUAL_ENV=/app/.venv
-ENV PATH="$VIRTUAL_ENV/bin:$PATH"
-
 # Copy source code
 COPY . .
 
-# Build Rust application with PyO3
+# Build Rust application
 RUN cargo build --release
 
-# Stage 3: Runtime image with Python 3.11 and compiled binary
-FROM python:3.11-slim-bookworm
+# Stage 2: Runtime image
+FROM debian:bookworm-slim
 
 # Install minimal runtime dependencies
 RUN apt-get update && apt-get install -y \
@@ -65,19 +35,12 @@ RUN apt-get update && apt-get install -y \
 
 WORKDIR /app
 
-# Copy Python virtual environment with all dependencies
-COPY --from=python-builder /app/.venv /app/.venv
-
 # Copy compiled Rust binary
 COPY --from=rust-builder /usr/src/app/target/release/oramacore /app/
 
 # Copy configuration
 COPY --from=rust-builder /usr/src/app/config.yaml /app/
 
-# Set Python environment variables for runtime
-ENV VIRTUAL_ENV=/app/.venv
-ENV PATH="$VIRTUAL_ENV/bin:$PATH"
-ENV PYO3_PYTHON=/app/.venv/bin/python3
 ENV RUST_LOG=oramacore=trace
 
 EXPOSE 8080
