--- requirements-cpu.txt
@@ -1,23 +1,0 @@
-# Python 3.11 suggested
-
-numpy==2.3.4
-
-python-dotenv==1.1.1
-fastembed==0.7.4
-
-torch==2.9.0
-torchvision==0.24.0
-torchaudio==2.9.0
-
-opentelemetry-api==1.29.0
-opentelemetry-sdk==1.29.0
-
-pyyaml==6.0.2
-cachetools==6.2.1
-json-repair==0.52.0
-
-onnxruntime==1.23.1
-accelerate==1.10.1
-einops==0.8.1
-
-pydantic==2.11.7
