--- src/python/scripts/pyproject.toml
@@ -1,2 +1,0 @@
-[tool.black]
-line-length = 120
